Modularisation du code et intégration des questions au module ampleur
